(#52) 데이터 클래스 수정 및 데이터 클래스 사용 코드 수정

+++ app/src/main/java/com/boostcamp/travery/data/model/UserAction.kt
@@ -25,8 +25,8 @@ data class UserAction(
     @ColumnInfo(name = "hashtag") var hashTag: String?,
     @ColumnInfo(name = "main_image") var mainImage: String?,
     @ColumnInfo(name = "sub_image") var subImage: String?,
-    @ColumnInfo(name = "latitude") var latitude: Long,
-    @ColumnInfo(name = "longitude") var longitude: Long,
+    @ColumnInfo(name = "latitude") var latitude: Double,
+    @ColumnInfo(name = "longitude") var longitude: Double,
     @ColumnInfo(name = "course_code", index = true) var courseCode: Long?
 ) : Parcelable {
     @PrimaryKey(autoGenerate = true)
